% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateMosaic.R
\name{bivariateMosaic}
\alias{bivariateMosaic}
\title{A bivariateMosaic Function}
\usage{
bivariateMosaic(
  dat,
  title = "Bivariate analysis",
  color = c("blue", "yellow", "red"),
  limit = c(0, 0.05)
)
}
\arguments{
\item{dat}{data results from BivariateAnalysis function}

\item{title}{graph title}

\item{color}{characters of color name for low, mid, and high values respectively.}

\item{limit}{scale of the heat plot. Two number to represent minimum, and maximum range to plot.}
}
\description{
This function produces a heatmap for the bivariate analysis result.
}
\examples{
data(iris)
result = bivariateAnalysis(iris)
#Plot only significant correlation/association
bivariateMosaic(result[result$pval <0.05,])
}
\references{
\url{https://www.tandfonline.com/doi/full/10.1080/10618600.2018.1473780}
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{bivariate}
\keyword{plot}
