% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidity.R
\name{checkValidity}
\alias{checkValidity}
\title{A checkValidity Function}
\usage{
checkValidity(
  dat,
  datatype,
  showOnlyInvalid = F,
  saveToFile = F,
  fileName = "datasummary"
)
}
\arguments{
\item{dat}{data in data frame format.}

\item{datatype}{characters of data types.}

\item{showOnlyInvalid}{TRUE if you want the function to return only invalid data.}

\item{saveToFile}{TRUE if you wish to save the data summary to a file.}

\item{fileName}{a file name used to save the data summary.}
}
\description{
This function checks that the designated data are valid given its data type. It also produces data summary
including data type and data range.
}
\examples{
library(FactoMineR)
data(wine)
datatype = c("factor", "numeric", "numeric", "numeric", "numeric", "integer", "numeric", "numeric",
"numeric", "numeric", "numeric", "numeric", "numeric", "integer")
checkValidity(wine, datatype)
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{checking}
\keyword{data}
