% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVariableNames.R
\name{checkVariableNames}
\alias{checkVariableNames}
\title{A checkVariableNames Function}
\usage{
checkVariableNames(raw, parameterName)
}
\arguments{
\item{raw}{Data in data.frame format.}

\item{parameterName}{Factor or characters vector of parameters' names.}
}
\description{
This function checks if the given parameters' names and the columns' names of the data are the same.
}
\examples{
data(iris)
checkVariableNames(iris,c("Sepal.Length","Sepal.Width","Petal.Length","Petal.Width","Species"))
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{check}
\keyword{name}
\keyword{variable}
