% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImputation.R
\name{dataImputation}
\alias{dataImputation}
\title{A dataImputation Function}
\usage{
dataImputation(dat)
}
\arguments{
\item{dat}{data to be imputed in data.frame format}
}
\value{
A list contains imputation models with imputed data, and a summary of imputation results.
}
\description{
This function performs impuation using 3 techniques i.e. Random Forest (RF), Multiple Imputation by
Chained Equations (MICE), and MICE with RF. The imputed data are check against the original data to see if
the imputed data distribution differs from original data distribution.
}
\examples{
library(FactoMineR)
data(wine)
dataImputation = function(wine)
}
\references{
\url{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-019-3110-0}

\url{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.169.5745&rep=rep1&type=pdf}
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{imputation}
