% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hClustCon.R
\name{hClustCon}
\alias{hClustCon}
\title{A clusterAnalysis Function}
\usage{
hClustCon(
  x,
  k = 1,
  diss = F,
  scale = T,
  metric = "euclidean",
  method = "ward",
  hierConsol = T,
  consolMet = "kmean",
  ...
)
}
\arguments{
\item{x}{a data frame or matrix of data}

\item{k}{integer number of clusters}

\item{diss}{If FALSE, then x is treated as a matrix of observations by variables. Otherwise, distant matrix should be given.}

\item{scale}{TRUE to scale and center data before clustering}

\item{metric}{character string specifying the metric to be used for calculating dissimilarities between observations.
The currently available options are "euclidean" and "manhattan". Euclidean distances are root sum-of-squares of
differences, and manhattan distances are the sum of absolute differences. If x is already a dissimilarity matrix,
then this argument will be ignored}

\item{method}{character string defining the hierarchical clustering method. See agnes for more details.}

\item{hierConsol}{TRUE to perform cluster consolidation.}

\item{consolMet}{clustering methods to use. The function implements 4 clustering algorithms: "kmean" (k-mean),
"fuzzy" (Fuzzy c-mean), "pam" (Partitioning Around Medoids), and "hierarchy" (Hierarchical Clustering).}

\item{...}{arguments pass on to hierarhical clustering algorithm functions.}
}
\value{
Return a list containing
\item{t}{Hierarchical tree object}
\item{cluster}{Cluster outcomes}
}
\description{
This function perform a hierarchical clustering, with an option of using other clustering methods e.g. k-mean,
fuzzy c-mean, or Partitioning Around Medoids to consolidate clustering results.
}
\examples{
data(iris)
hClustCon(iris[, -5])
}
\references{
\url{https://www.sciencedirect.com/topics/computer-science/clustering-technique}
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{analysis}
\keyword{clustering}
