% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingAnalysis.R
\name{missingAnalysis}
\alias{missingAnalysis}
\title{A missingAnalysis Function}
\usage{
missingAnalysis(dat, lower.limit = 0.5, plot = TRUE)
}
\arguments{
\item{dat}{data set in data frame format.}

\item{lower.limit}{the correlation limit that specifies significant pattern between variables.}

\item{plot}{TRUE to produce plots.}
}
\value{
Return a list containing missing data proportion and missingness correlation
}
\description{
This function performs missing data pattern Analysis. The function calculates correlation between missing data
between variables.
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{data}
\keyword{missing}
