% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeforeAfterImpute.R
\name{plotBeforeAfterImpute}
\alias{plotBeforeAfterImpute}
\title{A plotBeforeAfterImpute Function}
\usage{
plotBeforeAfterImpute(before, after, col.names = NULL)
}
\arguments{
\item{before}{data before imputation in data.frame format}

\item{after}{data fater imputation in data.frame format}

\item{col.names}{Name of column}
}
\description{
This function plots graphs to compare before and after data imputation. This function is used to
assess the quality of the imputed data.
}
\examples{
data(iris)
#create missing data
before = iris
before[sample(1:nrow(iris), round(0.1*nrow(iris))), 1] = NA
before[sample(1:nrow(iris), round(0.1*nrow(iris))), 3] = NA
after = missForest::missForest(before)$ximp
plotBeforeAfterImpute(before, after)
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{analysis}
\keyword{imputation}
