% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxplotGroup.R
\name{plotBoxplotGroup}
\alias{plotBoxplotGroup}
\title{A plotBoxplotGroup Function}
\usage{
plotBoxplotGroup(x, groups, scale = T, max.plot = 20, color.blind = TRUE)
}
\arguments{
\item{x}{numerical observations in data.frame format.}

\item{groups}{groups in factor.}

\item{scale}{a logical value to indicate if data centering and scaling should be used. It is recommend to use
this when producing a box plot.}

\item{max.plot}{maximium plot size}

\item{color.blind}{TRUE to use color pallett suitable for color blind}
}
\description{
This function plots box plots by groups.
}
\examples{
data(iris)
cl = kmeans(iris[, -5], 3)
plotBoxplotGroup(iris[, -5], cl$cluster)
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{box}
\keyword{plot}
