% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterProfile.R
\name{plotClusterProfile}
\alias{plotClusterProfile}
\title{A plotClusterProfile Function}
\usage{
plotClusterProfile(
  x,
  clusters,
  plot.type = "both",
  max.group = 4,
  col = c("green", "white", "red"),
  color.blind = F,
  stat = "median"
)
}
\arguments{
\item{x}{outcome observations in data.frame format}

\item{clusters}{clusters in factor format}

\item{plot.type}{Sepcify plot type for numerical data. Use "both" will produce both heatmap and boxplot, otherwise
specify either "boxplot" or "heatmap".}

\item{max.group}{number of graphs per plot}

\item{col}{characters specifying a set of colors used to represent low, mid, and high values, respectively in heatmap.}

\item{color.blind}{logical operator to indicate if color blind pallete should be used.}

\item{stat}{"mean" to use mean or "median" to use meidan when calculating average values}
}
\description{
This function plots clusters profile including a heatmap for numerical variables and
bar chart for categorical variables.
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{analysis}
\keyword{cluster}
