% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmapGroup.R
\name{plotHeatmapGroup}
\alias{plotHeatmapGroup}
\title{A plotHeatmapGroup Function}
\usage{
plotHeatmapGroup(dat)
}
\arguments{
\item{dat}{data in  data.frame format.}
}
\description{
This function plots a heatmap of a data frame
}
\examples{
data(iris)
plotHeatmapGroup(iris[,1:4])
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{heatmap}
