% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preCheck.R
\name{preCheck}
\alias{preCheck}
\title{A preCheck Function}
\usage{
preCheck(dataset, datatype)
}
\arguments{
\item{dataset}{data in data.frame.}

\item{datatype}{data.frame object containing information about variable names,    type,    whether the variable
should be excluded,    the source of the variable,    and whether the variable is a response/target.}
}
\description{
This function performs an initial checking for the given data set and data type. The checking process includes data format of dataset
and data type. Columns' names for the data type object and variables' names of the data set.
}
\examples{
data(iris)
datatype = data.frame("Parameter" = c("Sepal.Length","Sepal.Width","Petal.Length",
                      "Petal.Width","Species"),
                      "Type" = c("numeric","numeric","numeric","numeric","factor"),
                      "Excluded" = c(0,0,0,0,0),
                      "Source" = c(NA,NA,NA,NA,NA),
                      "Response" = c(0,0,0,0,1))
preCheck(iris, datatype)
}
\author{
Saisakul Chernbumroong, Henry  Nanji
}
\keyword{checking}
\keyword{data}
